requirements.txt
aiohttp
python-dotenv
fastapi
uvicorn
python-multipart
pandas
scikit-learn
nltk